requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Naked==0.1.31
paho-mqtt==1.4.0
pkg-resources==0.0.0
protobuf==3.6.1
pycodestyle==2.4.0
pycryptodome==3.7.2
pylint==2.2.2
PyYAML==3.13
requests==2.21.0
rope==0.11.0
shellescape==3.4.1
six==1.12.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.10.11
